Exercise 8.1.2
Fixed type error in Id setter

+++ src/main/java/com/codeup/models/Post.java
@@ -12,7 +12,7 @@ public class Post {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
